Create Favorite Product function and view
