+++ TechShopSolution.ApiIntegration/ICustomerApiClient.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using TechShopSolution.ViewModels.Catalog.Customer;
+using TechShopSolution.ViewModels.Catalog.Product;
 using TechShopSolution.ViewModels.Common;
 using TechShopSolution.ViewModels.Location;
 using TechShopSolution.ViewModels.Sales;
@@ -29,6 +30,8 @@ namespace TechShopSolution.ApiIntegration
         Task<ApiResult<OrderPublicViewModel>> GetOrderDetail(int id, int cus_id);
         Task<ApiResult<string>> ConfirmDoneShip(int id);
         Task<ApiResult<string>> CancelOrder(OrderCancelRequest request);
+        Task<ApiResult<bool>> RatingPoduct(ProductRatingRequest request);
+        Task<ApiResult<bool>> FavoriteProduct(int cus_id, int product_id);
 
     }
 }
