+++ TechShopSolution.ApiIntegration/IProductApiClient.cs
@@ -32,6 +32,5 @@ namespace TechShopSolution.ApiIntegration
         Task<PublicProductsViewModel> GetBestSellerProducts(int take);
         Task<PublicProductsViewModel> GetProductsByCategory(int id, int take);
         Task<List<ProductOverViewModel>> GetProductsRelated(int id, int take);
-        Task<ApiResult<bool>> RatingPoduct(ProductRatingRequest request);
     }
 }
