+++ TechShopSolution.ApiIntegration/ProductApiClient.cs
@@ -404,20 +404,6 @@ namespace TechShopSolution.ApiIntegration
                 return JsonConvert.DeserializeObject<List<ProductOverViewModel>>(body);
             return JsonConvert.DeserializeObject<List<ProductOverViewModel>>(body);
         }
-        public async Task<ApiResult<bool>> RatingPoduct(ProductRatingRequest request)
-        {
-            var client = _httpClientFactory.CreateClient();
-            client.BaseAddress = new Uri(_configuration["BaseAddress"]);
-
-            var json = JsonConvert.SerializeObject(request);
-            var httpContent = new StringContent(json, Encoding.UTF8, "application/json");
-
-            var respone = await client.PostAsync($"/api/product/rating", httpContent);
-            var result = await respone.Content.ReadAsStringAsync();
-            if (respone.IsSuccessStatusCode)
-                return JsonConvert.DeserializeObject<ApiSuccessResult<bool>>(result);
-            else return JsonConvert.DeserializeObject<ApiErrorResult<bool>>(result);
-        }
 
     }
 }
