+++ TechShopSolution.Application/Catalog/Customer/ICustomerService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 using TechShopSolution.ViewModels.Catalog.Customer;
+using TechShopSolution.ViewModels.Catalog.Product;
 using TechShopSolution.ViewModels.Common;
 using TechShopSolution.ViewModels.Location;
 using TechShopSolution.ViewModels.Sales;
@@ -22,5 +23,8 @@ namespace TechShopSolution.Application.Catalog.Customer
         Task<PagedResult<CustomerViewModel>> GetAllPaging(GetCustomerPagingRequest request);
         Task<bool> VerifyEmail(string email);
         List<OrderViewModel> GetLatestOrder(int id, int take);
+        Task<ApiResult<bool>> RatingProduct(ProductRatingRequest request);
+        Task<ApiResult<bool>> FavoriteProduct(int cus_id, int product_id);
+
     }
 }
