+++ TechShopSolution.BackendApi/Controllers/CustomerController.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using TechShopSolution.Application.Catalog.Customer;
 using TechShopSolution.ViewModels.Catalog.Customer;
+using TechShopSolution.ViewModels.Catalog.Product;
 
 namespace TechShopSolution.BackendApi.Controllers
 {
@@ -102,6 +103,22 @@ namespace TechShopSolution.BackendApi.Controllers
         public IActionResult GetLatestOrder(int id, int take)
         {
             var result = _customerService.GetLatestOrder(id, take);
+            return Ok(result);
+        }
+        [HttpPost("rating")]
+        public async Task<IActionResult> RatingProduct(ProductRatingRequest request)
+        {
+            var result = await _customerService.RatingProduct(request);
+            if (!result.IsSuccess)
+                return BadRequest(result);
+            return Ok(result);
+        }
+        [HttpGet("favorite")]
+        public async Task<IActionResult> FavoriteProduct(int cus_id, int product_id)
+        {
+            var result = await _customerService.FavoriteProduct(cus_id, product_id);
+            if (!result.IsSuccess)
+                return BadRequest(result);
             return Ok(result);
         }
     }
