+++ TechShopSolution.BackendApi/Controllers/ProductController.cs
@@ -150,13 +150,5 @@ namespace TechShopSolution.BackendApi.Controllers
                 return BadRequest(result);
             return Ok(result);
         }
-        [HttpPost("rating")]
-        public async Task<IActionResult> RatingProduct(ProductRatingRequest request)
-        {
-            var result = await _productService.SaveRating(request);
-            if (!result.IsSuccess)
-                return BadRequest(result);
-            return Ok(result);
-        }
     }
 }
