+++ TechShopSolution.WebApp/Views/Product/Detail.cshtml
@@ -1,20 +1,21 @@
 ﻿@model ProductDetailViewModel
 @inject Microsoft.Extensions.Configuration.IConfiguration Configuration;
 @using TechShopSolution.WebApp.Models
+@using System.Security.Claims;
 @{
     int score5 = 0;
-    int score4 = 0;
-    int score3 = 0;
-    int score2 = 0;
-    int score1 = 0;
-    bool flag = true;
+ int score4 = 0;
+ int score3 = 0;
+ int score2 = 0;
+ int score1 = 0;
+ bool flag = true;
 }
 @section FacebookComments{
     <div id="fb-root"></div>
     <script async defer crossorigin="anonymous" src="https://connect.facebook.net/vi_VN/sdk.js#xfbml=1&version=v11.0&appId=846669762547678&autoLogAppEvents=1" nonce="xvuFE9SJ"></script>
 }
 @section Script{
-    <script src="~/js/ProductRatingScoreBoard.js"></script>
+    <script src="~/js/Product.js"></script>
     <script src="~/js/OrderCancelReason.js"></script>
 }
 @if (Model == null)
@@ -134,31 +135,36 @@ else
                                     <div class="btnGroup">
                                         <button type="button" class="btn btn-large btn-primary btn-add-cart" data-id="@Model.Product.id"> Thêm vào <i class=" icon-shopping-cart"></i></button>
                                         <button type="button" class="btn btn-large btn-danger btn-buy-now" data-id="@Model.Product.id"> Mua ngay</button>
-                                        <div class="text-right margin-right-30">
-                                            <button class="btn">♡ Yêu thích</button>
-                                        </div>
                                     </div>
                                 }
                                 else
                                 {
                                     <span class="OutOfStock">Hết hàng</span>
-
-                                    <div class="text-right margin-right-30">
-                                        <button class="btn">♡ Yêu thích</button>
-                                    </div>
                                 }
                             }
+                            <div class="pull-right d-flex margin-right-20">
+                                <div class="product-client-infos-item">
+                                    <i class="icon-eye-open product-client-infos-icon"></i>&nbsp; @Model.Product.view_count
+                                </div>
+                                <div class="product-client-infos-item">
+                                    <i class="icon-heart-empty product-client-infos-icon"></i>&nbsp; <span id="favorite-total">125</span>
+                                </div>
+                            </div>
                         </div>
                     </div>
                 </form>
-                <div class="pull-right d-flex margin-right-20">
-                    <div class="product-client-infos-item">
-                        <i class="icon-eye-open product-client-infos-icon"></i>&nbsp; @Model.Product.view_count
+                @if (User.Identity.IsAuthenticated)
+                {
+                    <div class="text-right margin-right-30" id="favorite-button-area">
+                        <button class="btn" id="btnFavoriteProduct" data-cusid="@User.FindFirst(ClaimTypes.Sid).Value" data-productid="@Model.Product.id">♡ Yêu thích</button>
                     </div>
-                    <div class="product-client-infos-item">
-                        <i class="icon-heart-empty product-client-infos-icon"></i>&nbsp; 125
+                }
+                else
+                {
+                    <div class="text-right margin-right-30">
+                        <button class="btn" onclick="alert('Quý khách cần đăng nhập để sử dụng chức năng này')">♡ Yêu thích</button>
                     </div>
-                </div>
+                }
             </div>
             <div class="span9">
                 <a href="#" name="detail"></a>
