+++ TechShopSolution.WebApp/Views/Shared/_Layout.cshtml
@@ -47,7 +47,7 @@
     <div class="alert alert-success" id="snackbar" role="alert">
         <b class="ReultMessage"></b>
     </div>
-    <div class="alert alert-success" id="snackbarDanger" role="alert">
+    <div class="alert alert-danger" id="snackbarDanger" role="alert">
         <b id="ErrorMessage"></b>
     </div>
